requirements.txt
gunicorn==19.9.0
meinheld==0.6.1
alembic==1.0.0
flask==1.0.2
flask-restful==0.3.6
flask_sqlalchemy==2.3.2
flask-cors==3.0.6
flask_script==2.0.6
schematics==2.1.0
python-decouple==3.1
psycopg2==2.7.5
aiohttp==3.4.4
elastic-apm==4.0.1
pyjwt==1.7.0
kafka-python==1.4.4
psycopg2-binary==2.7.6.1
git+git://github.com/GymWorkoutApp/gwa-framework.git@0.0.36